Move dndState to common module It's in the right package now
