+++ common/build.gradle.kts
@@ -24,4 +24,5 @@ android {
 
 dependencies {
     implementation(libs.androidx.core.ktx)
+    implementation(libs.kotlinx.coroutines.core)
 }
