+++ mobile/src/main/kotlin/com/boswelja/smartwatchextensions/dndsync/Utils.kt
@@ -1,10 +1,7 @@
 package com.boswelja.smartwatchextensions.dndsync
 
 import android.app.NotificationManager
-import android.content.BroadcastReceiver
 import android.content.Context
-import android.content.Intent
-import android.content.IntentFilter
 import androidx.core.content.getSystemService
 import androidx.lifecycle.asFlow
 import com.boswelja.smartwatchextensions.common.dndsync.References.DND_STATUS_PATH
@@ -13,47 +10,10 @@ import com.boswelja.smartwatchextensions.common.preference.PreferenceKey.DND_SYN
 import com.boswelja.smartwatchextensions.common.toByteArray
 import com.boswelja.smartwatchextensions.watchmanager.WatchManager
 import java.util.UUID
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.channels.awaitClose
-import kotlinx.coroutines.channels.sendBlocking
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.callbackFlow
 import kotlinx.coroutines.flow.first
 import timber.log.Timber
 
 object Utils {
-
-    /**
-     * Gets a [Flow] of this watches DnD state.
-     */
-    @ExperimentalCoroutinesApi
-    fun Context.dndState(): Flow<Boolean> = callbackFlow {
-        val notificationManager = getSystemService<NotificationManager>()!!
-        val dndChangeReceiver = object : BroadcastReceiver() {
-            override fun onReceive(context: Context, intent: Intent?) {
-                if (intent?.action == NotificationManager.ACTION_INTERRUPTION_FILTER_CHANGED) {
-                    sendBlocking(notificationManager.isDndEnabled)
-                }
-            }
-        }
-        val filter = IntentFilter().apply {
-            addAction(NotificationManager.ACTION_INTERRUPTION_FILTER_CHANGED)
-        }
-        this@dndState.registerReceiver(dndChangeReceiver, filter)
-
-        send(notificationManager.isDndEnabled)
-
-        awaitClose {
-            this@dndState.unregisterReceiver(dndChangeReceiver)
-        }
-    }
-
-    /**
-     * Checks whether DnD is enabled for this watch.
-     * @return true if DnD is enabled, false otherwise.
-     */
-    private val NotificationManager.isDndEnabled: Boolean
-        get() = this.currentInterruptionFilter != NotificationManager.INTERRUPTION_FILTER_ALL
 
     suspend fun handleDnDStateChange(
         context: Context,
