+++ wearos/src/main/kotlin/com/boswelja/smartwatchextensions/dndsync/DnDLocalChangeListener.kt
@@ -16,9 +16,8 @@ import com.boswelja.smartwatchextensions.common.dndsync.References.DND_STATUS_PA
 import com.boswelja.smartwatchextensions.common.dndsync.References.DND_SYNC_LOCAL_NOTI_ID
 import com.boswelja.smartwatchextensions.common.dndsync.References.DND_SYNC_NOTI_CHANNEL_ID
 import com.boswelja.smartwatchextensions.common.dndsync.References.START_ACTIVITY_FROM_NOTI_ID
+import com.boswelja.smartwatchextensions.common.dndsync.dndState
 import com.boswelja.smartwatchextensions.common.toByteArray
-import com.boswelja.smartwatchextensions.dndsync.Observers.dndState
-import com.boswelja.smartwatchextensions.dndsync.Observers.theaterMode
 import com.boswelja.smartwatchextensions.extensions.extensionSettingsStore
 import com.boswelja.smartwatchextensions.main.ui.MainActivity
 import com.boswelja.smartwatchextensions.phoneStateStore
